+++ src/runtime/gc_ctrl.h
@@ -3,7 +3,7 @@
 
 #include "misc.h"
 
-extern long
+extern unsigned long long
      stat_minor_words,
      stat_promoted_words,
      stat_major_words,
@@ -11,7 +11,7 @@ extern long
      stat_major_collections,
      stat_heap_size;
 
-void init_gc (long, long, int, int);
+void init_gc (unsigned long long, unsigned long long, int, int);
 
 
 #endif /* _gc_ctrl_ */
